phase3-ecs/services/data-ingestion/requirements.txt
flask==3.0.0
boto3==1.34.34
psycopg2-binary==2.9.9
requests==2.31.0
gunicorn==21.2.0
